Refactor deleteById to throw if not found

+++ src/main/java/com/ghazal/spring6restmvc/service/CustomerServiceJpa.java
@@ -56,8 +56,12 @@ public class CustomerServiceJpa implements CustomerService {
     }
 
     @Override
-    public void deleteCustomerById(UUID id) {
-        customerRepository.deleteById(id);
+    public Boolean deleteCustomerById(UUID id) {
+        if (customerRepository.existsById(id)){
+            customerRepository.deleteById(id);
+            return true;
+        }
+        return false;
 
     }
